requirements.txt
autopep8==1.4.4
beautifulsoup4==4.8.1
certifi==2019.11.28
chardet==3.0.4
idna==2.8
pycodestyle==2.5.0
PyYAML==5.2
requests==2.22.0
soupsieve==1.9.5
urllib3==1.25.7
